% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch data from the Data Catalogue and return as one object}
\usage{
fetch_data(
  ckan_url = Sys.getenv("CKAN_URL", unset = "https://data.dbca.wa.gov.au")
)
}
\arguments{
\item{ckan_url}{The URL of the CKAN data catalogue to fetch data from.
Inferred from env var \code{CKAN_URL}, default: "https://data.dbca.wa.gov.au".}
}
\description{
Downloads all data files from CKAN and saves individual
tables as \code{fst} to \code{inst/data}. One data file can generate multiple tables.
}
\seealso{
Other extract: 
\code{\link{resume_data}()}
}
\concept{extract}
