% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume_data.R
\name{resume_data}
\alias{resume_data}
\title{Read data from locally stored snapshots.}
\usage{
resume_data()
}
\description{
Locally stored snapshots are generated by \code{fetch_data}.
}
\seealso{
Other extract: 
\code{\link{fetch_data}()}
}
\concept{extract}
