% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch data from the Data Catalogue and return as one object}
\usage{
fetch_data()
}
\description{
Fetch data from the Data Catalogue and return as one object
}
\concept{extract}
